requirements.txt
flask==3.1.2
flask-cors==6.0.2
rasterio==1.4.4
pillow==12.0.0
numpy==2.3.5
shapely==2.1.2
